requirements.txt
openai == 0.28
pandas
python-dotenv
Flask